requirements.txt
python-telegram-bot
python-dotenv
web3
ccxt
